pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-rl-endgames"
version = "0.1.0"
description = "Research-oriented reinforcement learning experiments on solved chess endgames."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-chess>=1.999",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
rl = [
  "torch>=2.2",
]
analysis = [
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "pandas>=2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
